#OSIS-2956 Résolution du probleme d'import cyclique
