+++ base/views/education_groups/perms.py
@@ -64,15 +64,6 @@ def can_delete_education_group(user, education_group):
     return True
 
 
-def can_delete_all_education_group(user, education_group):
-    pers = get_object_or_404(Person, user=user)
-    education_group_years = EducationGroupYear.objects.filter(education_group=education_group)
-    for education_group_yr in education_group_years:
-        if not business_perms.is_eligible_to_delete_education_group(pers, education_group_yr, raise_exception=True):
-            raise PermissionDenied
-    return True
-
-
 def can_change_general_information(view_func):
     def f_can_change_general_information(request, *args, **kwargs):
         person = get_object_or_404(Person, user=request.user)
